Rethink status table, some refactoring
