+++ constants.py
@@ -1,2 +1,8 @@
 
 STAR_DIRECTORY = "starsystems"
+PLANET_NOUNS =  ['planet', 'p', 'pl']
+
+DIPLOMACY = "diplomacy"
+STRENGTH = "strength"
+SCIENCE = "science"
+
